Use a decimal when possible
Before, when WroksheetReader was able to parse a decimal in GetNumericValue, it would be converted either into a single or a double. This is a problem when reading some values out of Excel, for instance 4372449.78 which would be converted to a Single and rounded to 4372450. This may optimize for performance, but that seems to be not really what is wanted here when reading out of Excel.

By just returning the decimal when possible, the highest possible precision is retained  when parsing numbers.